+++ example/lib/cupertino.dart
@@ -1,12 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:pick_country_picker/pick_country_picker.dart';
 
-void main() => runApp(ExampleApp());
+void main() => runApp(const ExampleApp());
 
 class ExampleApp extends StatelessWidget {
+  const ExampleApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Country Picker Example',
       home: HomePage(),
@@ -15,6 +17,8 @@ class ExampleApp extends StatelessWidget {
 }
 
 class HomePage extends StatefulWidget {
+  const HomePage({Key? key}) : super(key: key);
+
   @override
   _HomePageState createState() => _HomePageState();
 }
@@ -39,13 +43,13 @@ class _HomePageState extends State<HomePage> {
           hideSearch: false,
           selectedCountryIsoCode: selectedCountry?.iso2Code,
           title: 'Select your country',
-          priorityCountryCodes: ['US', 'CA', 'GB', 'LV'],
+          priorityCountryCodes: const ['US', 'CA', 'GB', 'LV'],
           onCountryChanged: (Country country) {
             setState(() => selectedCountry = country);
             Navigator.of(context).pop();
           },
           countryDisplayBuilder: (Country country) {
-            return '${country.countryName}';
+            return country.countryName;
           },
           flagBuilder: (Country country) {
             return Image.asset(
@@ -65,7 +69,7 @@ class _HomePageState extends State<HomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Country Picker Example'),
+        title: const Text('Country Picker Example'),
       ),
       body: Center(
         child: Column(
@@ -73,12 +77,12 @@ class _HomePageState extends State<HomePage> {
           children: <Widget>[
             ElevatedButton(
               onPressed: _showCountryPicker,
-              child: Text('Show Country Picker'),
+              child: const Text('Show Country Picker'),
             ),
-            SizedBox(height: 20),
+            const SizedBox(height: 20),
             if (selectedCountry != null) ...[
               Text('Selected Country: ${selectedCountry!.countryName}'),
-              SizedBox(height: 10),
+              const SizedBox(height: 10),
               Image.asset(
                 selectedCountry!.flagUri!,
                 package: 'pick_country_picker',